split out workflow functions
