+++ lib/proposals/workflows/getWorkflowTemplates.ts
@@ -1,0 +1,13 @@
+import { log } from '@charmverse/core/log';
+import type { Prisma } from '@charmverse/core/prisma-client';
+import { prisma } from '@charmverse/core/prisma-client';
+import type { ProposalWorkflowTyped } from '@charmverse/core/proposals';
+
+export async function getWorkflowTemplates(spaceId: string) {
+  const dbWorkflows = await prisma.proposalWorkflow.findMany({
+    where: {
+      spaceId
+    }
+  });
+  return dbWorkflows as ProposalWorkflowTyped[];
+}
